backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pypdf2==3.0.1
langchain==0.0.340
langchain-community==0.0.10
chromadb==0.4.22
openai==1.3.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
tiktoken==0.5.1
